feat: 도마뱀 현재 몸무게 저장 및 growth 테이블에 저장

+++ kiiun/src/main/java/com/cx_project/kiiun/domain/growth/entity/Growth.java
@@ -48,7 +48,7 @@ public class Growth {
     private LocalDateTime deletedAt;
 
     @Builder
-    public Growth(Long id, Integer weight, Lizard lizard, LocalDateTime createdAt, LocalDateTime modifiedAt, Boolean isDeleted, LocalDateTime deletedAt) {
+    public Growth(Long id, float weight, Lizard lizard, LocalDateTime createdAt, LocalDateTime modifiedAt, Boolean isDeleted, LocalDateTime deletedAt) {
         this.id = id;
         this.weight = weight;
         this.lizard = lizard;
